py-parsing-service/requirements.txt
fastapi
uvicorn[standard]
pymupdf
python-multipart
camelot-py[cv]
opencv-python
ghostscript
pytesseract
Pillow